% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_thresholds.R
\name{generate.null.bootstrap.matrix}
\alias{generate.null.bootstrap.matrix}
\title{Returns a matrix of parametric bootstrap samples from the null model of no locus effect}
\usage{
generate.null.bootstrap.matrix(scan.object, use.REML = TRUE, num.samples,
  seed = 1)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object.}

\item{use.REML}{DEFAULT: TRUE. Determines whether the variance components for the parametric sampling are 
based on maximizing the likelihood (ML) or the residual likelihood (REML).}

\item{num.samples}{The number of parametric bootstrap samples to return.}

\item{seed}{DEFAULT: 1. The sampling process is random, thus a seed must be set for samples to be the same
across machines.}
}
\description{
This function takes an scan.h2lmm() object, and returns a specified number of paratmetric bootstrap samples
from the null model of no locus effect.
}
\examples{
generate.null.bootstrap.matrix()
}
