% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_thresholds.R
\name{generate.perm.matrix}
\alias{generate.perm.matrix}
\title{Returns a matrix of (parametric) permutaions of the outcome}
\usage{
generate.perm.matrix(scan.object, use.REML = TRUE, num.samples, seed = 1)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object.}

\item{use.REML}{DEFAULT: TRUE. Determines whether the variance components for the parametric sampling are 
based on maximizing the likelihood (ML) or the residual likelihood (REML).}

\item{num.samples}{The number of permutation samples to return.}

\item{seed}{DEFAULT: 1. The sampling process is random, thus a seed must be set for samples to be consistent
across machines.}
}
\description{
This function takes an scan.h2lmm() object, and returns a specified number of permutation samples. When the
the null model contains covariates or there is a genetic relationship matrix, permutations are returned based
the maintains the structure of the data according to the null model.
}
\examples{
generate.perm.matrix()
}
