% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_bootstraps.R
\name{generate.positional.bootstrap.matrix}
\alias{generate.positional.bootstrap.matrix}
\title{Generate parametric bootstrap samples from the alternative model of a particular locus}
\usage{
generate.positional.bootstrap.matrix(scan.object.w.fit1, num.samples,
  seed = 1, use.BLUP = TRUE)
}
\arguments{
\item{scan.object.w.fit1}{A scan.h2lmm() object fit for a single locus. scan.h2lmm() includes the alternative
model fit when a single locus is specified. This alternative fit will be used as the basis for the bootstrap
sampling process.}

\item{num.samples}{The number of samples to be generated from the parametric model.}

\item{seed}{DEFAULT: 1. The sampling procedure is a random process. Specifying a seed allows consistent results
across runs and machines.}

\item{use.BLUP}{DEFAULT: TRUE. If a mixed effect model is specified with a random effect and corresponding kinship
matrix, the random effect could be sampled as well as the unstructured error term. If TRUE, the BLUP of the 
random effect is used, resulting in less random variation over samples, and ultimately narrower intervals.}
}
\description{
This function generates parametric bootstrap samples from the alternative model of a particular locus. These
samples will then be used to estimate a confidence interval for QTL position.
}
\examples{
generate.positional.bootstrap.matrix()
}
