% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_thresholds.R
\name{generate.sample.outcomes.matrix}
\alias{generate.sample.outcomes.matrix}
\title{Returns a matrix of outcome samples, either permutations or from the null model of no locus effect}
\usage{
generate.sample.outcomes.matrix(scan.object, model.type = c("null", "alt"),
  method = c("bootstrap", "permutation"), use.REML = TRUE,
  use.BLUP = FALSE, num.samples, seed = 1)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object.}

\item{method}{DEFAULT: "bootstrap". "bootstrap" specifies parametric bootstraps from the null model. "permutation" specifies
parametric permutations that respect the structure of the data. Permutations are more appropriate if the data have highly
influential data points.}

\item{use.REML}{DEFAULT: TRUE. Determines whether the variance components for the parametric sampling are 
based on maximizing the likelihood (ML) or the residual likelihood (REML).}

\item{num.samples}{The number of parametric bootstrap samples to return.}

\item{seed}{DEFAULT: 1. The sampling process is random, thus a seed must be set for samples to be consistent
across machines.}
}
\description{
This function takes an scan.h2lmm() object, and returns a specified number of outcome samples, either permutations or
from the null model of no locus effect.
}
\examples{
generate.sample.outcomes.matrix()
}
