% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instability_scans.R
\name{generate.simple.sample.outcomes.matrix}
\alias{generate.simple.sample.outcomes.matrix}
\title{Returns a matrix of outcome samples, either permutations or from the null model of no locus effect}
\usage{
generate.simple.sample.outcomes.matrix(formula, data,
  pheno.id = "SUBJECT.NAME", method = c("bootstrap", "permutation"),
  use.REML = TRUE, num.samples, seed = 1)
}
\arguments{
\item{method}{DEFAULT: "bootstrap". "bootstrap" specifies parametric bootstraps from the null model. "permutation" specifies
permutations that can respect the structure of the data. Permutations are more appropriate if the data have highly
influential data points.}

\item{use.REML}{DEFAULT: TRUE. Determines whether the variance components for the parametric sampling are 
based on maximizing the likelihood (ML) or the residual likelihood (REML).}

\item{num.samples}{The number of parametric bootstrap samples to return.}

\item{seed}{DEFAULT: 1. The sampling process is random, thus a seed must be set for samples to be consistent
across machines.}

\item{scan.object}{A scan.h2lmm() object.}
}
\description{
This function takes an scan.h2lmm() object, and returns a specified number of outcome samples, either permutations or
from the null model of no locus effect.
}
\examples{
generate.simple.sample.outcomes.matrix()
}
