% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.heatmap.R
\name{prob.heatmap}
\alias{prob.heatmap}
\title{Plot founder haplotype dosages/probabilities, ordered by phenotype}
\usage{
prob.heatmap(marker, p.value = NULL, genomecache, model = "additive",
  phenotype, phenotype.data, merge.by = "SUBJECT.NAME",
  founder.labels = NULL)
}
\arguments{
\item{marker}{A marker that is contained in the genome cache. In general this should be a marker of interest, such as one
beneath a putative QTL peak.}

\item{p.value}{DEFAULT: NULL. Includes the observed p-value in the plot title.}

\item{genomecache}{The path to the genome cache that contains founder haplotype information.}

\item{model}{DEFAULT: "additive". If "additive", dosages are plotted. If "full", probabilities are plotted.}

\item{phenotype}{The name of the phenotype column in the data set.}

\item{phenotype.data}{A data.frame object that contains the phenotype information. Should also have a column that matches
genomes in the genome cache.}

\item{merge.by}{DEFAULT: "SUBJECT.NAME". Specifies the columns to merge phenotype and haplotype data.}

\item{founder.labels}{DEFAULT: "NULL". If NULL, will default to the labels in the genome cache.}
}
\description{
This function produces a probability heatmap plot, ordered by the phenotype. This gives an idea of what the regression
procedure is actually being handed as inputs.
}
\examples{
prob.heatmap()
}
