% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{snp.genome.plotter.w.r2}
\alias{snp.genome.plotter.w.r2}
\title{Plot a single chromosome window of a SNP-based genome scan overlayed with r^2 information}
\usage{
snp.genome.plotter.w.r2(snp.scan, r2.object, scale = "Mb",
  y.max.manual = NULL, title = "", alt.col = NULL, this.cex = 1,
  hard.thresholds = NULL)
}
\arguments{
\item{snp.scan}{An imputed.snp.scan.h2lmm() object.}

\item{r2.object}{A pairwise.cor.snp.scan() object.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{title}{DEFAULT: "". Manually adds a max ylim to the plot. Allows multiple genome scans to easily be on the same scale.}

\item{alt.col}{DEFAULT: NULL. Allows for a custom color vector for individual SNPs.}

\item{this.cex}{DEFAULT: 1. Allows for the adjustment of the cex value for the main plot.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}
}
\description{
This function takes the outputs from imputed.snp.scan.h2lmm() and pairwise.cor.snp.scan() and plots 
the SNP association for a single chromosome, overlayed with r^2 information for a specified SNP.
}
\examples{
snp.genome.plotter.w.r2()
}
