% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{genome.plotter.whole}
\alias{genome.plotter.whole}
\title{Plot haplotype-based genome scan.}
\usage{
genome.plotter.whole(non.mi.scan.list, mi.scan = NULL, use.lod = TRUE,
  just.these.chr = NULL, scale = "Mb", main.colors = c("black", "gray48",
  "blue"), mi.colors = c("purple", "cyan", "dark green"), use.legend = TRUE,
  main = "", my.legend.cex = 0.6, bs.max = NULL, y.max.manual = NULL,
  hard.thresholds = NULL)
}
\arguments{
\item{non.mi.scan.list}{A list of scan.h2lmm() objects that are plotted within the same genome scan plot.}

\item{mi.scan}{DEFAULT: NULL. Multiple imputation scan.h2lmm() object. Median and confidence interval on median are plotted.}

\item{use.lod}{DEFAULT: TRUE. Plots either the LOD score or the -log10 p-value.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{main.colors}{DEFAULT: "black". The color of the main association score to be plotted.}

\item{use.legend}{DEFAULT: TRUE. Include a legend for the different associations. If TRUE, the labels are the names of the non.mi.scan.list object.}

\item{main}{DEFAULT: NULL. Adds a title above the model.}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}
}
\description{
This function takes the genome scan output from scan.h2lmm() and plots out genome scan.
}
